+++ ark_auth/src/server/smtp.rs
@@ -68,7 +68,7 @@ pub fn send_reset_password(
     let result = mailer
         .send(email.into())
         .map_err(|err| Error::Smtp(SmtpError::Lettre(err)))
-        .map(|res| ());
+        .map(|_res| ());
     mailer.close();
     result
 }